pyproject.toml
[project]
name = "chunking-token-benchmark"
version = "0.1.0"
description = "Context recall/precision benchmark for text chunking methods"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ai21>=3.0.1",
    "bittensor>=7.4.0",
    "boto3>=1.36.2",
    "chromadb>=0.6.3",
    "cohere>=5.13.8",
    "fuzzywuzzy>=0.18.0",
    "joblib>=1.4.2",
    "langchain-ai21>=1.0.1",
    "langchain-core>=0.3.30",
    "langchain-openai>=0.3.0",
    "matplotlib>=3.10.0",
    "nltk>=3.9.1",
    "openai>=1.59.8",
    "pandas>=2.2.3",
    "pinecone>=5.4.2",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.0.1",
    "ragas==0.2.11",
    "regex>=2024.11.6",
    "scikit-learn>=1.6.1",
    "tabulate>=0.9.0",
    "tiktoken>=0.8.0",
    "tqdm>=4.67.1",
    "unstructured-client>=0.29.0",
]
